Moved UpVar* types to a separate file.
